+++ lib/src/extension/string_extension.dart
@@ -3,6 +3,8 @@ extension SkadiStringExtension on String {
     return this[0].toUpperCase() + substring(1, length);
   }
 
+  bool get isNetworkUrl => startsWith("http") || startsWith("www");
+
   String get fileExtension {
     return split(".").last;
   }
@@ -10,9 +12,14 @@ extension SkadiStringExtension on String {
   num? get toNum {
     return num.tryParse(this);
   }
+
+  String emptyReplace(String replacement) {
+    return isEmpty ? replacement : this;
+  }
 }
 
 extension SkadiNullableStringExtension on String? {
   bool get isNullOrEmpty => this == null || (this?.isEmpty ?? true);
+
   bool get isNotNullOrEmpty => !isNullOrEmpty;
 }