+++ lib/src/utilities/types.dart
@@ -5,3 +5,5 @@ import 'package:flutter/material.dart';
 typedef CustomErrorWidget = Widget Function(dynamic, BuildContext);
 
 typedef FutureOrCallBack = FutureOr<void> Function();
+
+typedef JsonMap = Map<String, dynamic>;